chore: Move cli.h
So it can be used to parse compile_commands.json files.
